pyproject.toml
[project]
name = "vrchat-ime-chat"
version = "0.1.0"
description = "VRChat用のIME対応フローティングチャットウィンドウ"
authors = [
    { name = "⛐", email = "26d0@tuta.io" }
]
dependencies = [
    "customtkinter>=5.2.2",
    "python-osc>=1.9.3",
    "nuitka>=2.5.9",
    "pillow>=11.1.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "B",    # flake8-bugbear
    "I",    # isort
    "W",    # pycodestyle warnings
    "C90",  # mccabe complexity
    "N",    # pep8-naming
    "D",    # pydocstring
    "UP",   # pyupgrade
    "ANN",  # flake8-annotations
    "S",    # bandit
    "A",    # flake8-builtins
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.isort]
known-first-party = ["vrchat_ime_chat"]
force-single-line = true

[dependency-groups]
dev = [
    "go-task-bin>=3.41.0",
    "nuitka>=2.5.9",
]
